requirements.txt
argcomplete==1.9.3
arrow==0.12.1
boto3==1.9.60
botocore==1.12.60
certifi==2018.11.29
cfn-flip==1.1.0.post1
chardet==3.0.4
Click==7.0
docutils==0.14
durationpy==0.5
Flask==1.0.2
Flask-Cors==3.0.7
future==0.16.0
hjson==3.0.1
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.3
kappa==0.6.0
knackpy==0.0.12
lambda-packages==0.20.0
MarkupSafe==1.1.0
Pillow==5.3.0
placebo==0.8.2
python-dateutil==2.6.1
python-slugify==1.2.4
PyYAML==3.13
requests==2.20.1
s3transfer==0.1.13
six==1.11.0
toml==0.10.0
tqdm==4.19.1
troposphere==2.3.4
Unidecode==1.0.23
urllib3==1.24.1
Werkzeug==0.14.1
wsgi-request-logger==0.4.6
zappa==0.47.1
